moved extractor under framework to maintain localized dependencies only
